Re-arrange some classes and packages.
